Add support for append and write if it doesnt exist

+++ step_02_scrape_meta.rb
@@ -8,7 +8,7 @@ require './lib/utils.rb'
 # Config
 SLEEP_TIME           = 2
 
-raw_data = Utils::load_raw_data()
+raw_data = Utils::load_raw_data().reverse
 series_data = []
 existing_series = []
 
@@ -42,10 +42,17 @@ raw_data.each_with_index do |row, index|
   end
 end
 
-# Rewrite the file including the new meta informatio
-File.open(METADATA_OUTPUT, "w") do |file|
-  file.puts "Date;Title;URL;Source;Type;Runtime;Year;Rated;Released;Genre;Director;Actors;Rating;IMDB ID;Series Title;Triage"
-  raw_data.each do |row|
-   file.puts "#{row[:raw]};Netflix;#{row[:type]};#{row[:runtime]};#{row[:year]};#{row[:rated]};#{row[:released]};#{row[:genre]};#{row[:director]};#{row[:actors]};#{row[:rating]};#{row[:imdb_id]};#{row[:series_title]};#{row[:triage]}"
+if File.exist?(METADATA_OUTPUT)
+  File.open(METADATA_OUTPUT, "a") do |file|
+    raw_data.each do |row|
+     file.puts "#{row[:raw]};Netflix;#{row[:type]};#{row[:runtime]};#{row[:year]};#{row[:rated]};#{row[:released]};#{row[:genre]};#{row[:director]};#{row[:actors]};#{row[:rating]};#{row[:imdb_id]};#{row[:series_title]};#{row[:triage]}"
+    end
+  end
+else
+  File.open(METADATA_OUTPUT, "w") do |file|
+    file.puts "Date;Title;URL;Source;Type;Runtime;Year;Rated;Released;Genre;Director;Actors;Rating;IMDB ID;Series Title;Triage"
+    raw_data.each do |row|
+     file.puts "#{row[:raw]};Netflix;#{row[:type]};#{row[:runtime]};#{row[:year]};#{row[:rated]};#{row[:released]};#{row[:genre]};#{row[:director]};#{row[:actors]};#{row[:rating]};#{row[:imdb_id]};#{row[:series_title]};#{row[:triage]}"
+    end
   end
 end
